bir adjustments

+++ app/src/main/java/nerdvana/com/pointofsales/api_responses/AddProductToResponse.java
@@ -694,37 +694,37 @@ public class AddProductToResponse {
         private Object paymentTypeId;
         @SerializedName("rc")
         @Expose
-        private Integer rc;
+        private Double rc;
         @SerializedName("fnb")
         @Expose
-        private Integer fnb;
+        private Double fnb;
         @SerializedName("oth")
         @Expose
-        private Integer oth;
+        private Double oth;
         @SerializedName("ot_hours")
         @Expose
-        private Integer otHours;
+        private Double otHours;
         @SerializedName("ot_amount")
         @Expose
-        private Integer otAmount;
+        private Double otAmount;
         @SerializedName("xPerson_amount")
         @Expose
-        private Integer xPersonAmount;
+        private Double xPersonAmount;
         @SerializedName("discount")
         @Expose
-        private Integer discount;
+        private Double discount;
         @SerializedName("total")
         @Expose
-        private Integer total;
+        private Double total;
         @SerializedName("advance")
         @Expose
-        private Integer advance;
+        private Double advance;
         @SerializedName("tendered")
         @Expose
-        private Integer tendered;
+        private Double tendered;
         @SerializedName("change")
         @Expose
-        private Integer change;
+        private Double change;
         @SerializedName("is_soa")
         @Expose
         private Integer isSoa;
@@ -745,7 +745,7 @@ public class AddProductToResponse {
         private Double vat;
         @SerializedName("vat_exempt")
         @Expose
-        private Integer vatExempt;
+        private Double vatExempt;
         @SerializedName("xSkip")
         @Expose
         private Integer xSkip;
@@ -836,91 +836,91 @@ public class AddProductToResponse {
             this.paymentTypeId = paymentTypeId;
         }
 
-        public Integer getRc() {
+        public Double getRc() {
             return rc;
         }
 
-        public void setRc(Integer rc) {
+        public void setRc(Double rc) {
             this.rc = rc;
         }
 
-        public Integer getFnb() {
+        public Double getFnb() {
             return fnb;
         }
 
-        public void setFnb(Integer fnb) {
+        public void setFnb(Double fnb) {
             this.fnb = fnb;
         }
 
-        public Integer getOth() {
+        public Double getOth() {
             return oth;
         }
 
-        public void setOth(Integer oth) {
+        public void setOth(Double oth) {
             this.oth = oth;
         }
 
-        public Integer getOtHours() {
+        public Double getOtHours() {
             return otHours;
         }
 
-        public void setOtHours(Integer otHours) {
+        public void setOtHours(Double otHours) {
             this.otHours = otHours;
         }
 
-        public Integer getOtAmount() {
+        public Double getOtAmount() {
             return otAmount;
         }
 
-        public void setOtAmount(Integer otAmount) {
+        public void setOtAmount(Double otAmount) {
             this.otAmount = otAmount;
         }
 
-        public Integer getXPersonAmount() {
+        public Double getXPersonAmount() {
             return xPersonAmount;
         }
 
-        public void setXPersonAmount(Integer xPersonAmount) {
+        public void setXPersonAmount(Double xPersonAmount) {
             this.xPersonAmount = xPersonAmount;
         }
 
-        public Integer getDiscount() {
+        public Double getDiscount() {
             return discount;
         }
 
-        public void setDiscount(Integer discount) {
+        public void setDiscount(Double discount) {
             this.discount = discount;
         }
 
-        public Integer getTotal() {
+        public Double getTotal() {
             return total;
         }
 
-        public void setTotal(Integer total) {
+        public void setTotal(Double total) {
             this.total = total;
         }
 
-        public Integer getAdvance() {
+        public Double getAdvance() {
             return advance;
         }
 
-        public void setAdvance(Integer advance) {
+        public void setAdvance(Double advance) {
             this.advance = advance;
         }
 
-        public Integer getTendered() {
+        public Double getTendered() {
             return tendered;
         }
 
-        public void setTendered(Integer tendered) {
+        public void setTendered(Double tendered) {
             this.tendered = tendered;
         }
 
-        public Integer getChange() {
+        public Double getChange() {
             return change;
         }
 
-        public void setChange(Integer change) {
+        public void setChange(Double change) {
             this.change = change;
         }
 
@@ -972,11 +972,11 @@ public class AddProductToResponse {
             this.vat = vat;
         }
 
-        public Integer getVatExempt() {
+        public Double getVatExempt() {
             return vatExempt;
         }
 
-        public void setVatExempt(Integer vatExempt) {
+        public void setVatExempt(Double vatExempt) {
             this.vatExempt = vatExempt;
         }
